Push message

+++ app/src/main/java/badmitry/hellogeekbrains/model/Data.java
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName;
 import java.io.Serializable;
 
 public class Data implements Serializable {
-    @SerializedName("dt") public double dt;
+    @SerializedName("dt") public int dt;
     @SerializedName("wind") public Wind wind;
     @SerializedName("main") public Main main;
     @SerializedName("weather") public Weather[] weather;
